added bitrix support

+++ bitrix_main.py
@@ -1,6 +1,8 @@
 import asyncio
 import dataclasses
 
+from h2 import settings
+
 from database_connect_service.src import api
 from database_connect_service.src.site import get_enabled_api_settings, ApiSettings
 from database_connect_service.src.bitrix import get_unanswered_messages
@@ -48,7 +50,9 @@ async def qualification_execute(message, setting):
 
 async def process_bitrix(message, setting):
     print(message.bm.pipeline_id, message.bm.status_id)
-   #  get_btx_statuses_by_id(setting.id, setting.pipeline_id, setting.statuses_ids)
+    if not (message.bm.pipeline_id == setting.pipeline_id and message.bm.status_id in setting.statuses_ids):
+        return
+
     setting.prompt_context = setting.prompt_context.replace('\n', '    ')
     if message.message == 'restart':
         api.delete_messages(message.lead_id)
